Refactor Processor to use specialized mappers (SRP improvement)
